backend/pyproject.toml
[project]
name = "mcp-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asyncpg>=0.30.0",
    "certifi==2025.4.26",
    "click==8.2.0",
    "colorama==0.4.6",
    "databases[sqlite]>=0.9.0",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi==0.115.12",
    "fastapi-cli==0.0.7",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.16",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pydantic-settings>=2.9.1",
    "pygments==2.19.1",
    "python-dotenv==1.1.0",
    "python-jose>=3.4.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "rich==14.0.0",
    "rich-toolkit==0.14.5",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sqlalchemy>=2.0.40",
    "starlette==0.46.2",
    "typer==0.15.3",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "uvicorn==0.34.2",
    "watchfiles==1.0.5",
    "websockets==15.0.1",
    "yfinance>=0.2.60",
]
